THE SHIP SAILS: full player-move pipeline on original rules -- arrows -> classify -> arrive dispatch -> commit -> movement points
In-game self-test (headless): W into open water moves the caravel
(26,36)->(25,36) and deducts 12->9 movement thirds; N onto land with an empty
hold blocks silently (no lander aboard -- correct); SPACE refreshes movement
from the @UNIT column. Clean exit; audit 0 FAIL; frame-setup ALL PASS.

- src/platform/naval_glue.c: the dialog/commit leaves of naval_move_arrive:
  GAME.TXT @KEY message loader (prompt + options + @default), yes/no dialog
  (headless answers the section's own @default), the MOVE COMMIT
  (0x1A1F:0x142) as the byte-verified unlink+place pair with the thirds
  deduction (RECONSTRUCTED rule, cite-marked pending the order-engine
  decode), landfall begin/next = the chain walk, unload-at places the chosen
  lander at the mirrored dest, reveal/fortify-accum bound to their ported
  residents, power_scan_done/mark per their cites, sailhome/combat-fx logged
  stubs (later milestones).
- naval.c DS-word globals were declared `int` (4-byte reads on 2-byte DOS
  slots -- the 4083 disease); narrowed to int16_t and aliased byte-exactly
  (g_active_unit/0x5392, dest status/aux, power, europe flag, unit tables).
- Two more port diseases closed: func_005D84 (layer-[0x164] tile ADDRESS
  helper) routed to a writable host-layer address (the occupancy-nibble
  writer func_005E18 stores through it now), and its missing extern in the
  caller TU (implicit-int TRUNCATED the 64-bit pointer return -- the build's
  -Wno-implicit-function-declaration hides this whole hazard class).
- Shell: arrows call naval_move_arrive (which classifies and dispatches all
  nine statuses itself); SPACE = end-turn movement refresh; out-of-moves
  gate on the thirds counter.

+++ viceroy_source/src/combat/naval.c
@@ -49,19 +49,19 @@
 /* --- DGROUP globals (LOCAL externs — not promoted to globals.h) ----------- */
 /* [0x5392] active/selected unit index; *0x1C => UnitRecord. Written @0x6FBB.
  *          Read here @0x3FE08/@0x3FDE7 (imul ...,0x1c).                  [V]  */
-extern int      g_active_unit;          /* DS:0x5392 */
+extern int16_t  g_active_unit;          /* DS:0x5392 (word) */
 /* [0x9e4e] ship-destination STATUS code (1..9), produced by sibling
  *          func_03FA9C (sole writer @0x3FAA9/@0x3FAE4 etc.), consumed here
  *          @0x3FE16 (`MOV AX,[0x9e4e]`) as the dispatch selector.        [V]  */
-extern int      g_ship_dest_status;     /* DS:0x9e4e */
+extern int16_t  g_ship_dest_status;     /* DS:0x9e4e (word) */
 /* [0x9e50] secondary dest datum pushed to the LANDFALL handler @0x3FE56. [V]  */
-extern int      g_ship_dest_aux;        /* DS:0x9e50 */
+extern int16_t  g_ship_dest_aux;        /* DS:0x9e50 (word) */
 /* [0x5382] bit0 = "currently in Europe / cannot leave" flag, tested @0x3FEA6. */
 extern unsigned char g_in_europe_flag;  /* DS:0x5382 */
 /* [0x5394] active power index; *0x34 indexes a per-power tile-scan table at
  *          0x543e (+offset), where bit 0x80 = "scan done" — used by the
  *          landfall flood-scan loop @0x3FF81..0x3FFEF.                   [V]  */
-extern int      g_active_power;         /* DS:0x5394 */
+extern int16_t  g_active_power;         /* DS:0x5394 (word) */
 
 /* Per-unit-type table at DGROUP:0x5238 (stride 14): the high "movement/role"
  * byte compared to 0x63 (=99) @0x3FE95 to decide whether the just-revealed
